% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ModelCollection}
\alias{ModelCollection}
\title{Collection of pretrained models}
\usage{
ModelCollection(models)
}
\arguments{
\item{models}{A named or unnamed list of \code{\link{Model}} objects.}
}
\description{
An S7 class representing a collection of pretrained SigScore
\code{\link{Model}} objects. \code{ModelCollection} is used as a lightweight container
for grouping and managing multiple prediction models under a unified
interface.
}
\details{
All elements of \code{models} must inherit from the \code{\link{Model}} class.
Validation is performed at construction time to ensure type safety.
}
\section{Constructor}{

\code{ModelCollection(models)} creates a new collection of pretrained models.
}

\seealso{
\link{Model}, \code{\link[=sig_available_models]{sig_available_models()}}
}
