% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigscore.R
\name{sig_available_models}
\alias{sig_available_models}
\title{List available pretrained models}
\usage{
sig_available_models()
}
\value{
Invisibly returns a data frame describing available models.
}
\description{
Prints a table of all pretrained SigScore models available in the package.
}
